refactor(cli): extract gate parser owner
