+++ crates/tokmd/src/cli/parser.rs
@@ -24,6 +24,7 @@ mod analysis;
 mod cockpit;
 mod context;
 mod diff;
+mod gate;
 
 pub use analysis::{
     AnalysisPreset, CliAnalyzeArgs, EffortLayer, EffortModelKind, ImportGranularity, NearDupScope,
@@ -33,6 +34,7 @@ pub use context::{
     CliContextArgs, ContextOutput, ContextStrategy, HandoffArgs, HandoffPreset, ValueMetric,
 };
 pub use diff::{ColorMode, DiffArgs, DiffFormat};
+pub use gate::{CliGateArgs, GateFormat};
 
 #[derive(ValueEnum, Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(rename_all = "kebab-case")]
@@ -646,53 +648,6 @@ pub struct ToolsArgs {
     pub pretty: bool,
 }
 
-#[derive(Args, Debug, Clone)]
-pub struct CliGateArgs {
-    /// Input analysis receipt or path to scan.
-    #[arg(value_name = "INPUT")]
-    pub input: Option<PathBuf>,
-
-    /// Path to policy file (TOML format).
-    #[arg(long)]
-    pub policy: Option<PathBuf>,
-
-    /// Path to baseline receipt for ratchet comparison.
-    ///
-    /// When provided, gate will evaluate ratchet rules comparing current
-    /// metrics against the baseline values.
-    #[arg(long, value_name = "PATH")]
-    pub baseline: Option<PathBuf>,
-
-    /// Path to ratchet config file (TOML format).
-    ///
-    /// Defines rules for comparing current metrics against baseline.
-    /// Can also be specified inline in tokmd.toml under [[gate.ratchet]].
-    #[arg(long, value_name = "PATH")]
-    pub ratchet_config: Option<PathBuf>,
-
-    /// Analysis preset (for compute-then-gate mode).
-    #[arg(long, value_enum)]
-    pub preset: Option<AnalysisPreset>,
-
-    /// Output format.
-    #[arg(long, value_enum, default_value_t = GateFormat::Text)]
-    pub format: GateFormat,
-
-    /// Fail fast on first error.
-    #[arg(long)]
-    pub fail_fast: bool,
-}
-
-#[derive(ValueEnum, Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Default)]
-#[serde(rename_all = "kebab-case")]
-pub enum GateFormat {
-    /// Human-readable text output.
-    #[default]
-    Text,
-    /// JSON output.
-    Json,
-}
-
 #[derive(Args, Debug, Clone, Serialize, Deserialize)]
 pub struct SensorArgs {
     /// Base reference to compare from (default: main).
@@ -848,11 +803,6 @@ mod tests {
     #[test]
     fn context_output_default_is_list() {
         assert_eq!(ContextOutput::default(), ContextOutput::List);
-    }
-
-    #[test]
-    fn gate_format_default_is_text() {
-        assert_eq!(GateFormat::default(), GateFormat::Text);
     }
 
     #[test]
